order service changes

+++ order/src/main/java/dev/bhargav/ecommerce/dto/OrderLineResponse.java
@@ -1,8 +1,10 @@
 package dev.bhargav.ecommerce.dto;
 
+import java.math.BigDecimal;
+
 public record OrderLineResponse(
         Integer id,
-        double quantity
+        BigDecimal quantity
 ) {
 
 }
